Fix slicing.

+++ scripts/autokorrelation.py
@@ -16,7 +16,7 @@ ac = np.correlate(y, y, mode='full') / np.sum(y**2)
 # beidseitig bestimmt.  Im folgenden Schritt wird die erste Hälfte des
 # Vektors abgeschnitten.  (Sprich: "Von der Hälfte der Länge des Vektors
 # bis zu seinem Ende")
-ac = ac[ac.size / 2:]
+ac = ac[int(ac.size / 2):]
 
 # Plotten der Zeitreihe und der Autkokorrelation
 fig, ax = plt.subplots()
